phase-63: HS-63-02 MeetingSession mixins — four concerns out, the core at 795 lines
MeetingSession(TranscribeLoopMixin, IntelAnalysisMixin, PersistenceMixin,
MeetingMutationsMixin): the background transcription loop (270), the
live intel cadence (207), save() (145), and the action-item/title/tag
mutations (285) each live in their own module; session.py keeps
lifecycle, bookmarks, devices, and broadcasts at 795 lines (the
original file was 1,674). The verbatim proof is one line: the only
original body line lost is 'class MeetingSession:' itself, rewritten
as the composition. Zero test edits, exactly as the scaffold census
predicted.

Full suite 2768 passed, 17 skipped.
