+++ holdspeak/meeting_session/session.py
@@ -59,7 +59,18 @@ from .models import (
     _iso_or_none,
 )
 
-class MeetingSession:
+from .intel_analysis import IntelAnalysisMixin
+from .mutations import MeetingMutationsMixin
+from .persistence import PersistenceMixin
+from .transcribe_loop import TranscribeLoopMixin
+
+
+class MeetingSession(
+    TranscribeLoopMixin,
+    IntelAnalysisMixin,
+    PersistenceMixin,
+    MeetingMutationsMixin,
+):
     """Manages a background meeting recording session.
 
     Runs recording in background while allowing normal app operation.
@@ -762,234 +773,6 @@ class MeetingSession:
         except Exception as exc:
             log.error(f"Bookmark label generation failed: {exc}")
 
-    def update_action_item(self, item_id: str, status: str) -> Optional[dict]:
-        """Update the status of an action item.
-
-        Args:
-            item_id: The unique ID of the action item.
-            status: New status ("done", "pending", or "dismissed").
-
-        Returns:
-            The updated action item as a dict, or None if not found.
-        """
-        with self._lock:
-            if self._state is None or self._state.intel is None:
-                return None
-
-            normalized = str(status).strip().lower()
-            if normalized not in {"done", "pending", "dismissed"}:
-                log.warning(f"Unknown action item status: {status}")
-                return None
-
-            for item in self._state.intel.action_items:
-                if hasattr(item, "id") and item.id == item_id:
-                    if normalized == "done":
-                        item.mark_done()
-                    elif normalized == "dismissed":
-                        item.dismiss()
-                    else:
-                        item.status = "pending"
-                        item.completed_at = None
-                    log.info(f"Action item {item_id} updated to status={normalized}")
-                    return item.to_dict()
-
-                if isinstance(item, dict) and item.get("id") == item_id:
-                    item["status"] = normalized
-                    if normalized == "pending":
-                        item["completed_at"] = None
-                    else:
-                        item["completed_at"] = datetime.now().isoformat()
-                    log.info(f"Action item {item_id} updated to status={normalized}")
-                    return item
-
-            log.warning(f"Action item not found: {item_id}")
-            return None
-
-    def update_action_item_review(self, item_id: str, review_state: str) -> Optional[dict]:
-        """Update review state of an action item.
-
-        Args:
-            item_id: The unique ID of the action item.
-            review_state: New review state ("pending" or "accepted").
-
-        Returns:
-            Updated action item dict, or None if not found/invalid.
-        """
-        with self._lock:
-            if self._state is None or self._state.intel is None:
-                return None
-
-            normalized = str(review_state).strip().lower()
-            if normalized not in {"pending", "accepted"}:
-                log.warning(f"Unknown action item review_state: {review_state}")
-                return None
-
-            for item in self._state.intel.action_items:
-                if hasattr(item, "id") and item.id == item_id:
-                    if normalized == "accepted":
-                        if hasattr(item, "accept"):
-                            item.accept()
-                        else:
-                            item.review_state = "accepted"
-                            item.reviewed_at = datetime.now().isoformat()
-                    else:
-                        item.review_state = "pending"
-                        item.reviewed_at = None
-                    log.info(f"Action item {item_id} updated to review_state={normalized}")
-                    return item.to_dict()
-
-                if isinstance(item, dict) and item.get("id") == item_id:
-                    item["review_state"] = normalized
-                    item["reviewed_at"] = datetime.now().isoformat() if normalized == "accepted" else None
-                    log.info(f"Action item {item_id} updated to review_state={normalized}")
-                    return item
-
-            log.warning(f"Action item not found: {item_id}")
-            return None
-
-    def edit_action_item(
-        self,
-        item_id: str,
-        *,
-        task: str,
-        owner: Optional[str],
-        due: Optional[str],
-    ) -> Optional[dict]:
-        """Edit action-item content and mark it accepted.
-
-        Args:
-            item_id: The unique ID of the action item.
-            task: New task text.
-            owner: Optional owner text (None/empty clears).
-            due: Optional due text (None/empty clears).
-
-        Returns:
-            Updated action item dict, or None if not found/invalid.
-        """
-        clean_task = str(task).strip()
-        if not clean_task:
-            return None
-
-        clean_owner = owner.strip() if isinstance(owner, str) else None
-        clean_due = due.strip() if isinstance(due, str) else None
-
-        with self._lock:
-            if self._state is None or self._state.intel is None:
-                return None
-
-            for item in self._state.intel.action_items:
-                if hasattr(item, "id") and item.id == item_id:
-                    item.task = clean_task
-                    item.owner = clean_owner or None
-                    item.due = clean_due or None
-                    if hasattr(item, "accept"):
-                        item.accept()
-                    else:
-                        item.review_state = "accepted"
-                        item.reviewed_at = datetime.now().isoformat()
-                    log.info(f"Action item {item_id} edited and accepted")
-                    return item.to_dict()
-
-                if isinstance(item, dict) and item.get("id") == item_id:
-                    item["task"] = clean_task
-                    item["owner"] = clean_owner or None
-                    item["due"] = clean_due or None
-                    item["review_state"] = "accepted"
-                    item["reviewed_at"] = datetime.now().isoformat()
-                    log.info(f"Action item {item_id} edited and accepted")
-                    return item
-
-            log.warning(f"Action item not found: {item_id}")
-            return None
-
-    def set_title(self, title: str) -> None:
-        """Set meeting title.
-
-        Args:
-            title: The meeting title. Empty string clears the title.
-        """
-        with self._lock:
-            if self._state:
-                self._state.title = title.strip() or None
-                log.info(f"Meeting title set: {self._state.title}")
-                self._emit_broadcast("meeting_updated", {
-                    "title": self._state.title,
-                    "tags": self._state.tags,
-                })
-
-    def get_title(self) -> Optional[str]:
-        """Get current meeting title."""
-        with self._lock:
-            if self._state:
-                return self._state.title
-            return None
-
-    def add_tag(self, tag: str) -> bool:
-        """Add a tag to the meeting.
-
-        Args:
-            tag: Tag to add (will be lowercased and trimmed).
-
-        Returns:
-            True if tag was added, False if already exists or invalid.
-        """
-        with self._lock:
-            if self._state and tag.strip():
-                clean_tag = tag.strip().lower()
-                if clean_tag not in self._state.tags:
-                    self._state.tags.append(clean_tag)
-                    log.info(f"Tag added: {clean_tag}")
-                    self._emit_broadcast("meeting_updated", {
-                        "title": self._state.title,
-                        "tags": self._state.tags,
-                    })
-                    return True
-        return False
-
-    def remove_tag(self, tag: str) -> bool:
-        """Remove a tag from the meeting.
-
-        Args:
-            tag: Tag to remove.
-
-        Returns:
-            True if tag was removed, False if not found.
-        """
-        with self._lock:
-            if self._state:
-                clean_tag = tag.strip().lower()
-                if clean_tag in self._state.tags:
-                    self._state.tags.remove(clean_tag)
-                    log.info(f"Tag removed: {clean_tag}")
-                    self._emit_broadcast("meeting_updated", {
-                        "title": self._state.title,
-                        "tags": self._state.tags,
-                    })
-                    return True
-        return False
-
-    def set_tags(self, tags: list[str]) -> None:
-        """Replace all tags with a new list.
-
-        Args:
-            tags: List of tags (will be lowercased and trimmed).
-        """
-        with self._lock:
-            if self._state:
-                self._state.tags = [t.strip().lower() for t in tags if t.strip()]
-                log.info(f"Tags set: {self._state.tags}")
-                self._emit_broadcast("meeting_updated", {
-                    "title": self._state.title,
-                    "tags": self._state.tags,
-                })
-
-    def get_tags(self) -> list[str]:
-        """Get current meeting tags."""
-        with self._lock:
-            if self._state:
-                return list(self._state.tags)
-            return []
-
     def get_transcript(self) -> list[TranscriptSegment]:
         """Get all transcript segments."""
         with self._lock:
@@ -1010,451 +793,3 @@ class MeetingSession:
         if not segments:
             return ""
         return "\n".join(str(s) for s in segments)
-
-    def _transcribe_loop(self) -> None:
-        """Background thread that transcribes audio periodically."""
-        log.debug("Transcription loop started")
-
-        while not self._stop_event.is_set():
-            # Wait for interval or stop
-            self._stop_event.wait(self.TRANSCRIBE_INTERVAL)
-
-            if self._stop_event.is_set():
-                break
-
-            with self._lock:
-                if self._recorder is None:
-                    continue
-
-                # Get chunks since last transcription
-                mic_chunks, system_chunks = self._recorder.get_pending_chunks(
-                    since=self._last_transcribe_time
-                )
-                device_chunks = self._recorder.get_pending_device_chunks()
-
-            if mic_chunks or system_chunks or device_chunks:
-                self._transcribe_chunks(
-                    mic_chunks, system_chunks, device_chunks=device_chunks
-                )
-
-        log.debug("Transcription loop ended")
-
-    def _apply_overlap(self, stream_id: str, audio: "np.ndarray", final: bool) -> "np.ndarray":
-        """Prepend the previous pass's tail and stash a fresh tail for next.
-
-        AIPI-4-15 / HS-17 overlap windows. Returns the audio to feed
-        to Whisper (previous tail + current chunk). Updates the tail
-        store with the last ``self._overlap_tail_seconds`` of the
-        combined audio so next pass can re-prepend. On the final pass
-        the tail is cleared — no next pass to feed.
-        """
-        import numpy as _np
-
-        tail = self._stream_tails.get(stream_id)
-        if tail is not None and tail.size > 0:
-            audio = _np.concatenate([tail, audio])
-        if final:
-            self._stream_tails.pop(stream_id, None)
-        else:
-            tail_samples = int(self._overlap_tail_seconds * 16000)
-            self._stream_tails[stream_id] = (
-                audio[-tail_samples:].copy() if audio.size > tail_samples else audio.copy()
-            )
-        return audio
-
-    def _transcribe_chunks(
-        self,
-        mic_chunks: list[AudioChunk],
-        system_chunks: list[AudioChunk],
-        final: bool = False,
-        *,
-        device_chunks: Optional[dict[str, list[AudioChunk]]] = None,
-    ) -> None:
-        """Transcribe audio chunks and add to segments."""
-        new_segments: list[TranscriptSegment] = []
-        device_chunks = device_chunks or {}
-
-        # Process mic chunks
-        if mic_chunks:
-            mic_audio = concatenate_chunks(mic_chunks)
-            mic_audio = self._apply_overlap("mic", mic_audio, final)
-            duration = len(mic_audio) / 16000
-
-            if duration >= self.MIN_CHUNK_DURATION:
-                start_time = mic_chunks[0].timestamp
-                end_time = mic_chunks[-1].end_time
-
-                try:
-                    text = self.transcriber.transcribe(mic_audio)
-                    if text and text.strip():
-                        # Optionally diarize mic audio (for on-site meetings)
-                        speaker_id: Optional[str] = None
-                        speaker_name = self.mic_label
-                        if self.diarize_mic and self._diarizer is not None:
-                            try:
-                                speaker_id, speaker_name = self._diarizer.identify_speaker(mic_audio)
-                            except Exception as e:
-                                log.error(f"Mic speaker diarization error: {e}")
-
-                        segment = TranscriptSegment(
-                            text=text.strip(),
-                            speaker=speaker_name,
-                            speaker_id=speaker_id,
-                            start_time=start_time,
-                            end_time=end_time,
-                        )
-                        with self._lock:
-                            if self._state:
-                                self._state.segments.append(segment)
-                        new_segments.append(segment)
-                        if self.on_segment:
-                            try:
-                                self.on_segment(segment)
-                            except Exception as e:
-                                log.error(f"on_segment callback error: {e}")
-                        log.debug(f"Mic segment: {segment}")
-                except Exception as e:
-                    log.error(f"Mic transcription error: {e}")
-
-        # Process system chunks
-        if system_chunks:
-            system_audio = concatenate_chunks(system_chunks)
-            system_audio = self._apply_overlap("system", system_audio, final)
-            duration = len(system_audio) / 16000
-
-            if duration >= self.MIN_CHUNK_DURATION:
-                start_time = system_chunks[0].timestamp
-                end_time = system_chunks[-1].end_time
-
-                try:
-                    text = self.transcriber.transcribe(system_audio)
-                    if text and text.strip():
-                        # Identify speaker via diarization
-                        speaker_id: Optional[str] = None
-                        speaker_name = self.remote_label
-                        if self._diarizer is not None:
-                            try:
-                                speaker_id, speaker_name = self._diarizer.identify_speaker(system_audio)
-                            except Exception as e:
-                                log.error(f"Speaker diarization error: {e}")
-
-                        segment = TranscriptSegment(
-                            text=text.strip(),
-                            speaker=speaker_name,
-                            speaker_id=speaker_id,
-                            start_time=start_time,
-                            end_time=end_time,
-                        )
-                        with self._lock:
-                            if self._state:
-                                self._state.segments.append(segment)
-                        new_segments.append(segment)
-                        if self.on_segment:
-                            try:
-                                self.on_segment(segment)
-                            except Exception as e:
-                                log.error(f"on_segment callback error: {e}")
-                        log.debug(f"System segment: {segment}")
-                except Exception as e:
-                    log.error(f"System transcription error: {e}")
-
-        # Process device chunks (HS-14-06): each registered device's
-        # audio is transcribed independently, with the device's
-        # registered label as the speaker and its id stamped onto
-        # every produced segment.
-        for device_id, chunks in device_chunks.items():
-            if not chunks:
-                continue
-            audio = concatenate_chunks(chunks)
-            audio = self._apply_overlap(f"device:{device_id}", audio, final)
-            duration = len(audio) / 16000
-            if duration < self.MIN_CHUNK_DURATION:
-                continue
-
-            label = self.mic_label
-            if self._recorder is not None:
-                resolved_label = self._recorder.device_label(device_id)
-                if resolved_label:
-                    label = resolved_label
-
-            start_time = chunks[0].timestamp
-            end_time = chunks[-1].end_time
-            try:
-                text = self.transcriber.transcribe(audio)
-            except Exception as e:
-                log.error(f"Device {device_id!r} transcription error: {e}")
-                continue
-
-            if text and text.strip():
-                segment = TranscriptSegment(
-                    text=text.strip(),
-                    speaker=label,
-                    start_time=start_time,
-                    end_time=end_time,
-                    device_id=device_id,
-                )
-                with self._lock:
-                    if self._state:
-                        self._state.segments.append(segment)
-                new_segments.append(segment)
-                if self.on_segment:
-                    try:
-                        self.on_segment(segment)
-                    except Exception as e:
-                        log.error(f"on_segment callback error: {e}")
-                log.debug(f"Device segment ({device_id}): {segment}")
-
-        # Emit new segments to the observer's broadcast channel
-        for segment in new_segments:
-            self._emit_broadcast("segment", segment.to_dict())
-
-        # Update last transcribe time
-        if mic_chunks or system_chunks:
-            max_end = 0.0
-            if mic_chunks:
-                max_end = max(max_end, mic_chunks[-1].end_time)
-            if system_chunks:
-                max_end = max(max_end, system_chunks[-1].end_time)
-            self._last_transcribe_time = max_end
-
-        # Trigger intel analysis periodically
-        if new_segments:
-            self._segments_since_intel += len(new_segments)
-            if self._intel is not None and self._segments_since_intel >= self.INTEL_SEGMENT_INTERVAL:
-                self._maybe_run_intel()
-
-    def save(self, directory: Optional[Path] = None) -> MeetingSaveResult:
-        """Save meeting state to disk (JSON and SQLite database).
-
-        Args:
-            directory: Directory to save JSON to (default: ~/.local/share/holdspeak/meetings/).
-
-        Returns:
-            Structured save result covering both DB and JSON persistence.
-        """
-        with self._lock:
-            if self._state is None:
-                raise RuntimeError("No meeting to save")
-            state = self._state
-
-        database_saved = False
-        json_saved = False
-        database_error: Optional[str] = None
-        json_error: Optional[str] = None
-        json_path: Optional[Path] = None
-        intel_job_enqueued = False
-
-        # Save to SQLite database.
-        try:
-            from ..db import get_database
-
-            db = get_database()
-            db.meetings.save_meeting(state)
-            database_saved = True
-            if (
-                self.intel_enabled
-                and self.intel_deferred_enabled
-                and state.intel_status == "queued"
-                and state.segments
-            ):
-                db.intel.enqueue_intel_job(
-                    state.id,
-                    transcript_hash=state.transcript_hash(),
-                    reason=state.intel_status_detail,
-                )
-                intel_job_enqueued = True
-            log.info(f"Meeting saved to database: {state.id}")
-        except Exception as e:
-            database_error = f"{type(e).__name__}: {e}"
-            log.error(f"Failed to save meeting to database: {e}")
-
-        # HS-56-04: a wrapped meeting with open work / decisions is a moment
-        # the presence mascot reflects. Fires only for a finished meeting
-        # (ended_at set), only when the digest is non-empty, and never breaks
-        # the save.
-        if database_saved and state.ended_at is not None:
-            try:
-                from ..db import get_database as _get_db
-                from ..meeting_aftercare import build_aftercare_ready_event
-
-                event = build_aftercare_ready_event(_get_db(), state.id)
-                if event is not None:
-                    self._emit_broadcast("aftercare_ready", event)
-            except Exception as exc:  # observational only
-                log.debug(f"aftercare_ready broadcast skipped: {exc}")
-
-        # Save to JSON (backward compatibility).
-        if directory is None:
-            directory = Path.home() / ".local" / "share" / "holdspeak" / "meetings"
-
-        filename = f"meeting_{state.id}_{state.started_at.strftime('%Y%m%d_%H%M%S')}.json"
-        filepath = directory / filename
-
-        try:
-            directory.mkdir(parents=True, exist_ok=True)
-            with open(filepath, "w") as f:
-                json.dump(state.to_dict(), f, indent=2)
-            json_saved = True
-            json_path = filepath
-            log.info(f"Meeting saved to JSON: {filepath}")
-        except Exception as e:
-            json_error = f"{type(e).__name__}: {e}"
-            log.error(f"Failed to save meeting to JSON: {e}")
-
-        return MeetingSaveResult(
-            database_saved=database_saved,
-            json_saved=json_saved,
-            json_path=json_path,
-            database_error=database_error,
-            json_error=json_error,
-            intel_job_enqueued=intel_job_enqueued,
-        )
-
-    def _maybe_run_intel(self) -> None:
-        """Run intel analysis in background if not already running."""
-        if self._intel is None:
-            return
-
-        # Check if intel thread is already running
-        if self._intel_thread is not None and self._intel_thread.is_alive():
-            log.debug("Intel analysis already in progress, skipping")
-            return
-
-        self._segments_since_intel = 0
-        self._set_intel_status("running", "Analyzing the latest transcript window.")
-        self._intel_thread = threading.Thread(
-            target=self._run_intel_analysis,
-            daemon=True,
-        )
-        self._intel_thread.start()
-
-    def _run_intel_analysis(self, final: bool = False) -> None:
-        """Run intel analysis on current transcript with streaming support."""
-        if self._intel is None:
-            return
-
-        # Get current transcript
-        transcript = self.get_formatted_transcript()
-        if not transcript:
-            return
-
-        log.info(f"Running intel analysis (final={final}, streaming=True)")
-
-        # Generate a unique analysis ID to handle interruptions
-        analysis_id = str(uuid.uuid4())[:8]
-        self._current_analysis_id = analysis_id
-
-        try:
-            # Use streaming mode
-            stream_iter = self._intel.analyze(transcript, stream=True)
-
-            for chunk in stream_iter:
-                # Check if this analysis was interrupted by a newer one
-                if self._current_analysis_id != analysis_id:
-                    log.info(f"Intel analysis {analysis_id} interrupted by newer analysis")
-                    return
-
-                if isinstance(chunk, str):
-                    # Stream token to any observer (web dashboard)
-                    self._emit_broadcast("intel_token", chunk)
-                else:
-                    # Final IntelResult
-                    result = chunk
-                    if getattr(result, "error", None):
-                        detail = f"Deferred intel required: {result.error}"
-                        log.warning(f"Intel analysis deferred: {result.error}")
-                        if self.intel_deferred_enabled:
-                            self._deferred_intel_reason = result.error
-                            with self._lock:
-                                self._intel = None
-                            self._set_intel_status("queued", detail)
-                        else:
-                            self._set_intel_status("error", result.error, completed_at=datetime.now())
-                        return
-
-                    # Create snapshot - preserve ActionItem objects for status tracking
-                    snapshot = IntelSnapshot(
-                        timestamp=self.duration,
-                        topics=result.topics,
-                        action_items=result.action_items,  # Keep ActionItem objects
-                        summary=result.summary,
-                    )
-
-                    # Update state
-                    with self._lock:
-                        if self._state:
-                            self._state.intel = snapshot
-                            self._state.intel_status = "ready"
-                            self._state.intel_status_detail = (
-                                "Meeting intelligence ready."
-                                if not final
-                                else "Final meeting intelligence ready."
-                            )
-                            self._state.intel_completed_at = datetime.now()
-
-                    # Emit completion to any observer (web dashboard)
-                    self._emit_broadcast("intel_complete", snapshot.to_dict())
-                    self._emit_broadcast(
-                        "intel_status",
-                        self._get_state_dict().get("intel_status", {}),
-                    )
-
-                    # Callback
-                    if self.on_intel:
-                        try:
-                            self.on_intel(snapshot)
-                        except Exception as e:
-                            log.error(f"on_intel callback error: {e}")
-
-                    log.info(f"Intel analysis complete: {len(snapshot.topics)} topics, {len(snapshot.action_items)} action items")
-
-                    # Refine bookmark labels with full meeting context (final pass only)
-                    if final:
-                        self._refine_bookmark_labels(snapshot.summary)
-
-        except Exception as e:
-            log.error(f"Intel analysis failed: {e}")
-            if self.intel_deferred_enabled:
-                self._deferred_intel_reason = str(e)
-                with self._lock:
-                    self._intel = None
-                self._set_intel_status("queued", f"Deferred intel required: {e}")
-            else:
-                self._set_intel_status("error", str(e), completed_at=datetime.now())
-
-    def _refine_bookmark_labels(self, meeting_summary: str) -> None:
-        """Refine all bookmark labels using full meeting context.
-
-        Called during final analysis to improve bookmark labels with:
-        - High-level meeting summary for grounding
-        - Local ±10s context around each bookmark
-        """
-        if self._intel is None or self._state is None:
-            return
-
-        bookmarks = self._state.bookmarks
-        if not bookmarks:
-            return
-
-        log.info(f"Refining {len(bookmarks)} bookmark labels with meeting context")
-
-        for bookmark in bookmarks:
-            try:
-                # Get local context around bookmark
-                local_context = self._state.get_context_around(bookmark.timestamp, window=10.0)
-                if not local_context:
-                    continue  # No transcript near this bookmark
-
-                # Generate refined label with meeting context
-                label = self._intel.generate_bookmark_label_with_context(
-                    local_context=local_context,
-                    meeting_summary=meeting_summary,
-                )
-                if label and label != bookmark.label:
-                    old_label = bookmark.label
-                    with self._lock:
-                        bookmark.label = label
-                    log.info(f"Refined bookmark: '{old_label}' -> '{label}'")
-
-            except Exception as e:
-                log.error(f"Failed to refine bookmark at {bookmark.timestamp:.1f}s: {e}")
