+++ pm/roadmap/holdspeak/phase-63-backend-decomposition/evidence-story-02.md
@@ -1,0 +1,41 @@
+# Evidence — HS-63-02: MeetingSession mixins
+
+**Date:** 2026-06-12
+**Verdict:** done. `session.py` went from 1,460 to **795 lines**; the four
+concerns live in single-purpose mixin modules; **zero test edits**; the
+verbatim proof is one line — the class statement itself.
+
+## What shipped
+
+`holdspeak/meeting_session/` gained four mixins, composed by
+`MeetingSession(TranscribeLoopMixin, IntelAnalysisMixin, PersistenceMixin,
+MeetingMutationsMixin)`:
+
+- `transcribe_loop.py` (270): the background loop, the overlap window,
+  chunk transcription (incl. diarization handoff).
+- `intel_analysis.py` (207): the should-run cadence, the analysis pass,
+  bookmark-label refinement.
+- `persistence.py` (145): `save()` — the DB + JSON write and the
+  deferred-intel enqueue.
+- `mutations.py` (285): action-item status/review/edit, title, tags.
+
+`session.py` (795) keeps what the story scoped to it: `__init__`,
+lifecycle (start/stop), bookmarks, device attach/detach, broadcasts, the
+intel-status setters (used by lifecycle), and the read-only getters.
+Every mixin carries the same guarded optional-import header as the core
+(the HS-63-01 lesson applied: parent-relative from the start, all
+indentation levels).
+
+## The verbatim proof
+
+The body-line diff (imports/blanks excluded) between the ORIGINAL
+pre-phase `meeting_session.py` and the union of all six package files:
+**exactly one original line lost — `class MeetingSession:`**, rewritten
+as the mixin composition. Every method body is verbatim.
+
+## Proof
+
+- MRO smoke: the composed class resolves all moved methods.
+- Full suite: **2768 passed, 17 skipped** — zero test files touched
+  (this module has no monkeypatched globals; the census predicted zero
+  edits and zero happened).
