+++ pm/roadmap/holdspeak/phase-63-backend-decomposition/story-02-session-mixins.md
@@ -2,7 +2,7 @@
 
 - **Project:** holdspeak
 - **Phase:** 63
-- **Status:** backlog
+- **Status:** done
 - **Depends on:** HS-63-01
 - **Unblocks:** HS-63-05
 - **Owner:** unassigned
@@ -12,8 +12,8 @@ MeetingSession mixes the transcribe loop, intel analysis, persistence,
 and item/title/tag mutations with its lifecycle in one 1,400-line class.
 
 ## Scope
-- **In:** `holdspeak/meeting/` gains `transcribe_loop.py` (the loop +
-  overlap + chunk transcription), `intel.py` (the intel cadence +
+- **In:** `holdspeak/meeting_session/` (the HS-63-01 package) gains `transcribe_loop.py` (the loop +
+  overlap + chunk transcription), `intel_analysis.py` (the intel cadence +
   analysis + bookmark refinement), `persistence.py` (save), and
   `mutations.py` (action items, title, tags) as mixin classes with
   verbatim bodies; `meeting_session.py` keeps __init__, lifecycle
@@ -22,11 +22,13 @@ and item/title/tag mutations with its lifecycle in one 1,400-line class.
 - **Out:** WebRuntime (HS-63-03/04); behavior changes.
 
 ## Acceptance criteria
-- [ ] Each mixin is single-concern and under budget; bodies verbatim
+- [x] Each mixin is single-concern and under budget; bodies verbatim
       (locks travel with their methods).
-- [ ] `meeting_session.py` is the thin lifecycle + assembly module.
-- [ ] ZERO test edits (no monkeypatched globals exist on this module);
+- [x] `meeting_session.py` is the thin lifecycle + assembly module.
+- [x] ZERO test edits (no monkeypatched globals exist on this module);
       full suite green.
+
+      See `evidence-story-02.md`.
 
 ## Test plan
 - The full suite; the meeting/intel/diarization slices read specifically.
